requirements.txt
django
django-environ
django-bootstrap4
psycopg2-binary
gunicorn
google-cloud-secret-manager
google-auth
django-storages[google]